requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
requests
python-multipart
